feat: move more commons components to ui module
