+++ features/ui-common/build.gradle.kts
@@ -7,6 +7,7 @@ android {
 }
 
 dependencies {
+    implementation("com.wire.kalium:kalium-logic")
     implementation(libs.androidx.core)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
@@ -19,6 +20,8 @@ dependencies {
     implementation(libs.compose.material.core)
     implementation(libs.compose.material3)
     implementation(libs.compose.navigation)
+    implementation(libs.compose.material.icons)
+    implementation(libs.compose.ui.preview)
 
     implementation(libs.accompanist.systemUI)
     implementation(libs.visibilityModifiers)
