+++ features/ui-common/src/main/kotlin/com/wire/android/model/Clickable.kt
