+++ features/ui-common/src/main/kotlin/com/wire/android/ui/common/button/WireSecondaryIconButton.kt
@@ -35,7 +35,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import com.wire.android.R
 import com.wire.android.model.ClickBlockParams
 import com.wire.android.ui.common.dimensions
 import com.wire.android.ui.theme.wireDimensions
@@ -82,15 +81,25 @@ fun WireSecondaryIconButton(
 @Preview
 @Composable
 fun PreviewWireSecondaryIconButton() {
-    WireSecondaryIconButton({}, false, R.drawable.ic_add, 0)
+    WireSecondaryIconButton({}, false, com.google.android.material.R.drawable.m3_password_eye, 0)
 }
+
 @Preview
 @Composable
 fun PreviewWireSecondaryIconButtonLoading() {
-    WireSecondaryIconButton({}, true, R.drawable.ic_add, 0)
+    WireSecondaryIconButton({}, true, com.google.android.material.R.drawable.m3_password_eye, 0)
 }
+
 @Preview
 @Composable
 fun PreviewWireSecondaryIconButtonRound() {
-    WireSecondaryIconButton({}, false, R.drawable.ic_add, 0, CircleShape, DpSize(40.dp, 40.dp), DpSize(48.dp, 48.dp))
+    WireSecondaryIconButton(
+        {},
+        false,
+        com.google.android.material.R.drawable.m3_password_eye,
+        0,
+        CircleShape,
+        DpSize(40.dp, 40.dp),
+        DpSize(48.dp, 48.dp)
+    )
 }
